Create DirectEventSubmitter for web
